backend/app/services/link_scraper.py
import httpx
from bs4 import BeautifulSoup
from typing import Dict, Optional
from app.core.logging import get_logger

logger = get_logger(__name__)

async def scrape_github(github_url: str) -> Optional[Dict]:
    """Scrape GitHub profile for repositories and activity"""
    try:
        async with httpx.AsyncClient() as client:
            response = await client.get(github_url, follow_redirects=True, timeout=10.0)
            
            if response.status_code != 200:
                logger.warning(f"Failed to scrape GitHub: {github_url}")
                return None
            
            soup = BeautifulSoup(response.text, 'html.parser')
            
            # Extract basic info (this is simplified - real implementation would be more robust)
            data = {
                "url": github_url,
                "scraped": True,
                "repositories": [],  # Would extract repo list
                "contributions": 0,   # Would extract contribution count
                "followers": 0,       # Would extract followers
            }
            
            logger.info(f"Successfully scraped GitHub: {github_url}")
            return data
    
    except Exception as e:
        logger.error(f"Error scraping GitHub {github_url}: {str(e)}")
        return None

async def scrape_linkedin(linkedin_url: str) -> Optional[Dict]:
    """Scrape LinkedIn profile (note: LinkedIn has scraping restrictions)"""
    try:
        # Note: LinkedIn actively prevents scraping
        # In production, use LinkedIn API or manual data entry
        logger.warning("LinkedIn scraping is restricted. Using placeholder data.")
        
        data = {
            "url": linkedin_url,
            "scraped": False,
            "note": "LinkedIn API integration required for real data"
        }
        
        return data
    
    except Exception as e:
        logger.error(f"Error scraping LinkedIn {linkedin_url}: {str(e)}")
        return None

async def scrape_portfolio(portfolio_url: str) -> Optional[Dict]:
    """Scrape portfolio website for projects and information"""
    try:
        async with httpx.AsyncClient() as client:
            response = await client.get(portfolio_url, follow_redirects=True, timeout=10.0)
            
            if response.status_code != 200:
                logger.warning(f"Failed to scrape portfolio: {portfolio_url}")
                return None
            
            soup = BeautifulSoup(response.text, 'html.parser')
            
            # Extract basic metadata
            data = {
                "url": portfolio_url,
                "scraped": True,
                "title": soup.title.string if soup.title else "",
                "description": "",  # Would extract meta description
            }
            
            logger.info(f"Successfully scraped portfolio: {portfolio_url}")
            return data
    
    except Exception as e:
        logger.error(f"Error scraping portfolio {portfolio_url}: {str(e)}")
        return None

async def scrape_links(links: Dict[str, str]) -> Dict[str, Optional[Dict]]:
    """
    Scrape all provided links and return enriched data.
    
    Args:
        links: Dictionary with keys like 'github', 'linkedin', 'portfolio'
    
    Returns:
        Dictionary with scraped data for each link
    """
    enriched_data = {}
    
    if "github" in links:
        enriched_data["github"] = await scrape_github(links["github"])
    
    if "linkedin" in links:
        enriched_data["linkedin"] = await scrape_linkedin(links["linkedin"])
    
    if "portfolio" in links:
        enriched_data["portfolio"] = await scrape_portfolio(links["portfolio"])
    
    return enriched_data

